renamed probnum.type to probnum.typing
