+++ src/probnum/diffeq/odesolution.py
@@ -11,7 +11,7 @@ import numpy as np
 
 from probnum import _randomvariablelist, filtsmooth, randvars
 from probnum.filtsmooth._timeseriesposterior import DenseOutputLocationArgType
-from probnum.type import FloatArgType, IntArgType, RandomStateArgType, ShapeArgType
+from probnum.typing import FloatArgType, IntArgType, RandomStateArgType, ShapeArgType
 
 
 class ODESolution(filtsmooth.TimeSeriesPosterior):
