+++ src/probnum/kernels/_linear.py
@@ -5,7 +5,7 @@ from typing import Optional
 import numpy as np
 
 import probnum.utils as _utils
-from probnum.type import IntArgType, ScalarArgType
+from probnum.typing import IntArgType, ScalarArgType
 
 from ._kernel import Kernel
