+++ src/probnum/kernels/_rational_quadratic.py
@@ -6,7 +6,7 @@ import numpy as np
 import scipy.spatial.distance
 
 import probnum.utils as _utils
-from probnum.type import IntArgType, ScalarArgType
+from probnum.typing import IntArgType, ScalarArgType
 
 from ._kernel import Kernel
