+++ src/probnum/linops/_kronecker.py
@@ -7,7 +7,7 @@ from typing import Optional, Union
 
 import numpy as np
 
-from probnum.type import DTypeArgType
+from probnum.typing import DTypeArgType
 
 from . import _linear_operator, _utils
