+++ src/probnum/randprocs/_gaussian_process.py
@@ -5,7 +5,7 @@ from typing import Callable, Optional, Type, Union
 import numpy as np
 
 from probnum import kernels, randvars
-from probnum.type import RandomStateArgType, ShapeArgType
+from probnum.typing import RandomStateArgType, ShapeArgType
 
 from . import _random_process
