+++ src/probnum/randprocs/_random_process.py
@@ -7,7 +7,7 @@ import numpy as np
 
 from probnum import _randomvariablelist, randvars
 from probnum import utils as _utils
-from probnum.type import DTypeArgType, IntArgType, RandomStateArgType, ShapeArgType
+from probnum.typing import DTypeArgType, IntArgType, RandomStateArgType, ShapeArgType
 
 _InputType = TypeVar("InputType")
 _OutputType = TypeVar("OutputType")
