+++ tests/test_quad/test_kernel_embeddings.py
@@ -7,7 +7,7 @@ import pytest
 from scipy.linalg import sqrtm
 from scipy.special import roots_legendre
 
-from probnum.type import FloatArgType, IntArgType
+from probnum.typing import FloatArgType, IntArgType
 
 
 # Auxiliary functions
